Fix issue when using dictionary function
Previously after the function was evaluated the rest of the loop merging
the dicts containing the descriptions continued using V instead of the
return value of the function.

That could probably cause multiple issues, but the one guaranteed issue
was when the returned dictionary contained only plain descriptions and
no function (plain string instead of list).

If you are using a dictionary function and see
single-letter-descriptions where the letter is the second letter of the
description, this probably fixes it.